requirements.txt
flask
flask_restful
flask_jsonpify
gunicorn
werkzeug
tika
bs4
nltk
python-magic
numpy
tqdm
symspellpy==6.7.0
pandas==1.2.4
mistune==2.0.3
lxml==4.9.1
unidecode
nlm-utils
google-cloud-documentai
google-cloud-documentai-toolbox
termcolor
